Updated the DAL- need to implement funcs that the business layer will contact with

+++ dev/Inventory/src/DTO/ProductDTO.java
@@ -17,7 +17,7 @@ public class ProductDTO {
     private List<ItemDTO> itemDTOs;
 
     public ProductDTO(int id, String name, String manufacturer, int minCapacity, int buyingPrice, int sellingPrice,
-                      int weight, int inventoryCapacity, int storeCapacity, List<String> categories, List<ItemDTO> itemDTOs) {
+                      double weight, int inventoryCapacity, int storeCapacity, List<String> categories, List<ItemDTO> itemDTOs) {
         this.id = id;
         this.name = name;
         this.manufacturer = manufacturer;
